// mailmerge/src/engines/types.ts
export type TemplateEngineOptions = Record<string | number, unknown>;
// ...
export type TemplatePreview = {
    /**
     * Name of the preview, for you to recognise on re-render/load-back-in.
     *
     * This should be related to its content, and _should not be unique_ - we will append a unique identifier to the filename based on the CSV record.
     *
     * As this will be the last part of the filename, you should include a file extension
     * @example
     * "markdown-preview.md" // good
     * "preview.html" // good
     * `preview-${record["name"]}.html` // bad! We will add record specific parts to the filename
     */
    name: string;
    /** The content of the preview, to be written to a file */
    content: string;
    /** Metadata about the preview, to be written to a JSON file, that will be given back to you on re-render/load-back-in */
    metadata: Record<string, unknown>;
};

export type TemplatePreviews = TemplatePreview[];